requirements.txt
Django>=2.2.4
json-logic
psycopg2
boto3
gunicorn
django-storages
django-background-tasks